+++ src/GlobalServer/ICommandRunner.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace GlobalServer
+{
+    public interface ICommandRunner : IDisposable
+    {
+        Task Run();
+    }
+}
