+++ src/GlobalServer/NullCommand.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace GlobalServer
+{
+    public class NullCommand : ICommandRunner
+    {
+        public Task Run() 
+            => Task.CompletedTask;
+    }
+}
